webapp_on_linux/requirements.txt
flask
azure-storage
line-bot-sdk
pymongo